+++ h2/connection.py
@@ -19,7 +19,7 @@ from hpack.hpack import Encoder, Decoder
 from hpack.exceptions import HPACKError
 
 from .config import H2Configuration
-from .errors import PROTOCOL_ERROR, REFUSED_STREAM
+from .errors import ErrorCodes, _error_code_from_int
 from .events import (
     WindowUpdated, RemoteSettingsChanged, PingAcknowledged,
     SettingsAcknowledged, ConnectionTerminated, PriorityUpdated,
@@ -991,7 +991,8 @@ class H2Connection(object):
         :param stream_id: The ID of the stream to reset.
         :type stream_id: ``int``
         :param error_code: (optional) The error code to use to reset the
-            stream. Defaults to :data:`NO_ERROR <h2.errors.NO_ERROR>`.
+            stream. Defaults to :data:`ErrorCodes.NO_ERROR
+            <h2.errors.ErrorCodes.NO_ERROR>`.
         :type error_code: ``int``
         :returns: Nothing
         """
@@ -1379,7 +1380,7 @@ class H2Connection(object):
             for frame in self.incoming_buffer:
                 events.extend(self._receive_frame(frame))
         except InvalidPaddingError:
-            self._terminate_connection(PROTOCOL_ERROR)
+            self._terminate_connection(ErrorCodes.PROTOCOL_ERROR)
             raise ProtocolError("Received frame with invalid padding.")
         except ProtocolError as e:
             # For whatever reason, receiving the frame caused a protocol error.
@@ -1509,7 +1510,7 @@ class H2Connection(object):
             # remote peer now believes exists.
             if frame.stream_id in self._reset_streams:
                 f = RstStreamFrame(frame.promised_stream_id)
-                f.error_code = REFUSED_STREAM
+                f.error_code = ErrorCodes.REFUSED_STREAM
                 return [f], events
 
             raise ProtocolError("Attempted to push on closed stream.")
@@ -1713,7 +1714,7 @@ class H2Connection(object):
 
         # Fire an appropriate ConnectionTerminated event.
         new_event = ConnectionTerminated()
-        new_event.error_code = frame.error_code
+        new_event.error_code = _error_code_from_int(frame.error_code)
         new_event.last_stream_id = frame.last_stream_id
         new_event.additional_data = (frame.additional_data
                                      if frame.additional_data else None)