+++ h2/events.py
@@ -278,7 +278,7 @@ class RemoteSettingsChanged(object):
     When this event is received, the caller should confirm that the new
     settings are acceptable. If they are not acceptable, the user should close
     the connection with the error code :data:`PROTOCOL_ERROR
-    <h2.errors.PROTOCOL_ERROR>`.
+    <h2.errors.ErrorCodes.PROTOCOL_ERROR>`.
 
     .. versionchanged:: 2.0.0
        Prior to this version the user needed to acknowledge settings changes.
@@ -359,7 +359,8 @@ class StreamReset(object):
         #: The Stream ID of the stream that was reset.
         self.stream_id = None
 
-        #: The error code given.
+        #: The error code given. Either one of :class:`ErrorCodes
+        #: <h2.errors.ErrorCodes>` or ``int``
         self.error_code = None
 
         #: Whether the remote peer sent a RST_STREAM or we did.
@@ -463,7 +464,7 @@ class ConnectionTerminated(object):
     """
     def __init__(self):
         #: The error code cited when tearing down the connection. Should be
-        #: one of :data:`H2ERRORS <h2.errors.H2_ERRORS>`, but may not be if
+        #: one of :class:`ErrorCodes <h2.errors.ErrorCodes>`, but may not be if
         #: unknown HTTP/2 extensions are being used.
         self.error_code = None