+++ test/test_invalid_content_lengths.py
@@ -59,7 +59,7 @@ class TestInvalidContentLengths(object):
 
         expected_frame = frame_factory.build_goaway_frame(
             last_stream_id=1,
-            error_code=h2.errors.PROTOCOL_ERROR,
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR,
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -94,7 +94,7 @@ class TestInvalidContentLengths(object):
 
         expected_frame = frame_factory.build_goaway_frame(
             last_stream_id=1,
-            error_code=h2.errors.PROTOCOL_ERROR,
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR,
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -129,6 +129,6 @@ class TestInvalidContentLengths(object):
 
         expected_frame = frame_factory.build_goaway_frame(
             last_stream_id=1,
-            error_code=h2.errors.PROTOCOL_ERROR,
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR,
         )
         assert c.data_to_send() == expected_frame.serialize()