+++ test/test_invalid_headers.py
@@ -69,7 +69,7 @@ class TestInvalidFrameSequences(object):
             c.receive_data(data)
 
         expected_frame = frame_factory.build_goaway_frame(
-            last_stream_id=1, error_code=h2.errors.PROTOCOL_ERROR
+            last_stream_id=1, error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -308,7 +308,7 @@ class TestOversizedHeaders(object):
             c.receive_data(data)
 
         expected_frame = frame_factory.build_goaway_frame(
-            last_stream_id=1, error_code=h2.errors.ENHANCE_YOUR_CALM
+            last_stream_id=1, error_code=h2.errors.ErrorCodes.ENHANCE_YOUR_CALM
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -343,7 +343,7 @@ class TestOversizedHeaders(object):
             c.receive_data(data)
 
         expected_frame = frame_factory.build_goaway_frame(
-            last_stream_id=0, error_code=h2.errors.ENHANCE_YOUR_CALM
+            last_stream_id=0, error_code=h2.errors.ErrorCodes.ENHANCE_YOUR_CALM
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -377,7 +377,7 @@ class TestOversizedHeaders(object):
             c.receive_data(data)
 
         expected_frame = frame_factory.build_goaway_frame(
-            last_stream_id=0, error_code=h2.errors.ENHANCE_YOUR_CALM
+            last_stream_id=0, error_code=h2.errors.ErrorCodes.ENHANCE_YOUR_CALM
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -425,6 +425,6 @@ class TestOversizedHeaders(object):
             c.receive_data(data)
 
         expected_frame = frame_factory.build_goaway_frame(
-            last_stream_id=3, error_code=h2.errors.ENHANCE_YOUR_CALM
+            last_stream_id=3, error_code=h2.errors.ErrorCodes.ENHANCE_YOUR_CALM
         )
         assert c.data_to_send() == expected_frame.serialize()